requirements.txt
streamlit
pandas
python-dotenv
gspread
google-auth
gspread-dataframe
serpapi
groq
